fix some bugs in rewrite
